Move TestObserver to common module
