+++ stream-chat-android-test/src/main/java/io/getstream/chat/android/test/TestObserver.kt
@@ -1,16 +1,16 @@
-package com.getstream.sdk.chat.utils.livedata
+package io.getstream.chat.android.test
 
 import androidx.lifecycle.Observer
 
-internal class TestObserver<T> : Observer<T> {
-    var lastObservedValue: T? = null
+public class TestObserver<T> : Observer<T> {
+    public var lastObservedValue: T? = null
         private set
 
     override fun onChanged(value: T?) {
         lastObservedValue = value
     }
 
-    fun reset() {
+    public fun reset() {
         lastObservedValue = null
     }
 }
